market-ingestion/pyproject.toml
[project]
name = "market-ingestion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.134.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "supabase>=2.0.0",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
]

